+++ ModelingWebAppQuasar/quasar.conf.js
@@ -28,7 +28,6 @@ module.exports = configure(function (ctx) {
       'composition-api',
       'serverConnection',
       'i18n',
-      'firebase'
     ],
 
     // https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-css
