+++ ModelingWebAppQuasar/src/App.vue
@@ -9,16 +9,4 @@ import { defineComponent } from '@vue/composition-api';
 export default defineComponent({
   name: 'App',
 });
-
-</script>
-<script>
-import { mapActions } from "vuex";
-export default {
-  methods:{
-    ...mapActions('auth', ['handleAuthStateChange'])
-  },
-  mounted(){
-    this.handleAuthStateChange()
-  }
-}
 </script>
