+++ ModelingWebAppQuasar/src/class-diagram/src/class/class.rect.js
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/restrict-plus-operands */
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/restrict-plus-operands */
 import { Node, Rect } from 'topology-core'
 
 export function simpleClassIconRect(node) {
