+++ ModelingWebAppQuasar/src/router/routes.ts
@@ -1,13 +1,17 @@
 import { RouteConfig } from 'vue-router';
 const routes: RouteConfig[] = [
   {
-    path: '',
+    path: '/',
     component: () => import('layouts/MainLayout.vue'),
     children: [
-      { path: '/home', component: () => import('pages/Index.vue') },
-      { path: '/editor', component: () => import('components/Editor.vue') },
-      { path: '/signin', component: () => import('components/signin.vue') },
-      { path: '/signup', component: () => import('components/signup.vue') }
+      { path: '', component: () => import('pages/Index.vue') }
+    ]
+  },
+  {
+    path: '/editor',
+    component: () => import('layouts/MainLayout.vue'),
+    children: [
+      { path: '', component: () => import('components/Editor.vue') }
     ]
   },
   // Always leave this as last one,
