+++ ModelingWebAppQuasar/src/services/canvas.ts
@@ -1,0 +1,259 @@
+import { register as registerFlow } from 'topology-flow-diagram'
+import { register as registerActivity } from 'topology-activity-diagram'
+import { register as registerSequence } from 'topology-sequence-diagram'
+
+import { register as registerClass } from '../class-diagram/index'
+
+export function canvasRegister() {
+  registerFlow()
+  registerActivity()
+  registerSequence()
+  registerClass()
+}
+
+export const Tools = [
+  {
+    group: 'Diagrammes',
+    children: [
+      {
+        name: '开始',
+        icon: 'icon-inital',
+        data: {
+          text: '',
+          rect: {
+            width: 30,
+            height: 30
+          },
+          name: 'circle',
+          fillStyle: '#555',
+          strokeStyle: 'transparent'
+        }
+      },
+      {
+        name: '结束',
+        icon: 'icon-final',
+        data: {
+          text: '',
+          rect: {
+            width: 30,
+            height: 30
+          },
+          name: 'activityFinal'
+        }
+      },
+      {
+        name: '活动',
+        icon: 'icon-action',
+        data: {
+          text: '活动',
+          rect: {
+            width: 120,
+            height: 50
+          },
+          borderRadius: 0.25,
+          name: 'rectangle'
+        }
+      },
+      {
+        name: '决策/合并',
+        icon: 'icon-diamond',
+        data: {
+          text: '决策',
+          rect: {
+            width: 120,
+            height: 50
+          },
+          name: 'diamond'
+        }
+      },
+      {
+        name: '垂直泳道',
+        icon: 'icon-swimlane-v',
+        data: {
+          text: '垂直泳道',
+          rect: {
+            width: 200,
+            height: 500
+          },
+          name: 'swimlaneV'
+        }
+      },
+      {
+        name: '水平泳道',
+        icon: 'icon-swimlane-h',
+        data: {
+          text: '水平泳道',
+          rect: {
+            width: 500,
+            height: 200
+          },
+          name: 'swimlaneH'
+        }
+      },
+      {
+        name: '垂直分岔/汇合',
+        icon: 'icon-fork-v',
+        data: {
+          text: '',
+          rect: {
+            width: 10,
+            height: 150
+          },
+          name: 'forkV',
+          fillStyle: '#555',
+          strokeStyle: 'transparent'
+        }
+      },
+      {
+        name: '水平分岔/汇合',
+        icon: 'icon-fork',
+        data: {
+          text: '',
+          rect: {
+            width: 150,
+            height: 10
+          },
+          name: 'forkH',
+          fillStyle: '#555',
+          strokeStyle: 'transparent'
+        }
+      }
+    ]
+  },
+  {
+    group: '时序图和类图',
+    children: [
+      {
+        name: '生命线',
+        icon: 'icon-lifeline',
+        data: {
+          text: '生命线',
+          rect: {
+            width: 150,
+            height: 400
+          },
+          name: 'lifeline'
+        }
+      },
+      {
+        name: '激活',
+        icon: 'icon-focus',
+        data: {
+          text: '',
+          rect: {
+            width: 12,
+            height: 200
+          },
+          name: 'sequenceFocus'
+        }
+      },
+      {
+        name: '简单类',
+        icon: 'icon-simple-class',
+        data: {
+          text: 'Topolgoy',
+          rect: {
+            width: 270,
+            height: 200
+          },
+          paddingTop: 40,
+          font: {
+            fontFamily: 'Arial',
+            color: '#222',
+            fontWeight: 'bold'
+          },
+          fillStyle: '#ffffba',
+          strokeStyle: '#7e1212',
+          name: 'simpleClass',
+          children: [
+            {
+              text: '- name: string\n+ setName(name: string): void',
+              name: 'text',
+              paddingLeft: 10,
+              paddingRight: 10,
+              paddingTop: 10,
+              paddingBottom: 10,
+              rectInParent: {
+                x: 0,
+                y: 0,
+                width: '100%',
+                height: '100%',
+                rotate: 0
+              },
+              font: {
+                fontFamily: 'Arial',
+                color: '#222',
+                textAlign: 'left',
+                textBaseline: 'top'
+              }
+            }
+          ]
+        }
+      },
+      {
+        name: '类',
+        icon: 'icon-class',
+        data: {
+          text: 'Topolgoy',
+          rect: {
+            width: 270,
+            height: 200
+          },
+          paddingTop: 40,
+          font: {
+            fontFamily: 'Arial',
+            color: '#222',
+            fontWeight: 'bold'
+          },
+          fillStyle: '#ffffba',
+          strokeStyle: '#7e1212',
+          name: 'interfaceClass',
+          children: [
+            {
+              text: '- name: string',
+              name: 'text',
+              paddingLeft: 10,
+              paddingRight: 10,
+              paddingTop: 10,
+              paddingBottom: 10,
+              rectInParent: {
+                x: 0,
+                y: 0,
+                width: '100%',
+                height: '50%',
+                rotate: 0
+              },
+              font: {
+                fontFamily: 'Arial',
+                color: '#222',
+                textAlign: 'left',
+                textBaseline: 'top'
+              }
+            },
+            {
+              text: '+ setName(name: string): void',
+              name: 'text',
+              paddingLeft: 10,
+              paddingRight: 10,
+              paddingTop: 10,
+              paddingBottom: 10,
+              rectInParent: {
+                x: 0,
+                y: '50%',
+                width: '100%',
+                height: '50%',
+                rotate: 0
+              },
+              font: {
+                fontFamily: 'Arial',
+                color: '#222',
+                textAlign: 'left',
+                textBaseline: 'top'
+              }
+            }
+          ]
+        }
+      }
+    ]
+  }
+]
