--- ModelingWebAppQuasar/src/store/event.js
@@ -1,14 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-export const state = () => ({
-  event: {
-    name: '',
-    data: null
-  }
-})
-
-export const mutations = {
-  emit(state, event) {
-    state.event = event
-  }
-}
